Drop extra files
