+++ L1TMuon/interface/EMTFHit.h
@@ -1,162 +1,85 @@
-// Class for input trigger primitives to EMTF - AWB 04.01.16
-// Based on L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitive.h
-// In particular, see struct CSCData
+#ifndef __L1TMuon_EMTFHitExtra_h__
+#define __L1TMuon_EMTFHitExtra_h__
 
-#ifndef __L1TMuon_EMTFHit_h__
-#define __L1TMuon_EMTFHit_h__
-
+#include <cstdint>
 #include <vector>
-#include <boost/cstdint.hpp> 
-#include <cmath>
-#include <iostream>
-
-#include "DataFormats/MuonDetId/interface/CSCDetId.h"
-#include "DataFormats/MuonDetId/interface/RPCDetId.h"
-#include "DataFormats/CSCDigi/interface/CSCCorrelatedLCTDigi.h"
-#include "DataFormats/RPCDigi/interface/RPCDigi.h"
-//#include "DataFormats/L1TMuon/interface/EMTF/ME.h"
 
 namespace L1TMuonEndCap {
-  
-  class EMTFHit {
-  public:
-    
-  EMTFHit() :
-    
-    // Using -999 instead of -99 b/c this seems most common in the emulator.  Unfortunate. - AWB 17.03.16
-    endcap(-999), station(-999), ring(-999), sector(-999), sector_index(-999), subsector(-999), 
-      chamber(-999), csc_ID(-999), roll(-999), rpc_layer(-999), neighbor(-999), mpc_link(-999), 
-      wire(-999), strip(-999), strip_hi(-999), strip_low(-999), track_num(-999), quality(-999), 
-      pattern(-999), bend(-999), valid(-999), sync_err(-999), bc0(-999), bx(-999), stub_num(-999), 
-      is_CSC_hit(-999), is_RPC_hit(-999)
-      {};
-    
-    virtual ~EMTFHit() {};
 
-    //void ImportCSCDetId (const CSCDetId& _detId);
-    //CSCDetId CreateCSCDetId();
-    //void ImportRPCDetId (const RPCDetId& _detId);
-    //RPCDetId CreateRPCDetId();
-    //void ImportCSCCorrelatedLCTDigi (const CSCCorrelatedLCTDigi& _digi);
-    //CSCCorrelatedLCTDigi CreateCSCCorrelatedLCTDigi();
-    //void ImportRPCDigi (const RPCDigi& _digi);
-    //RPCDigi CreateRPCDigi();
-    //void ImportME (const emtf::ME _ME );
+  struct EMTFHitExtra {
 
-    //void PrintSimulatorHeader();
-    //void PrintForSimulator();
+    EMTFHitExtra() : detId(0),
+                     endcap(0), station(0), ring(0), roll(0), chamber(0), sector(0), subsector(0), csc_ID(0), cscn_ID(0),
+                     bx(0), subsystem(0),
+                     pc_sector(0), pc_station(0), pc_chamber(0), pc_segment(0),
+                     valid(0), strip(0), strip_low(0), strip_hi(0), wire(0), quality(0), pattern(0), bend(0),
+                     phi_fp(0), theta_fp(0), phzvl(0), ph_hit(0), zone_hit(0), zone_code(0),
+                     fs_segment(0), fs_zone_code(0), bt_station(0), bt_segment(0),
+                     phi_loc_deg(0), phi_glob_deg(0), theta_deg(0), eta(0)
+                     {}
 
-    void SetCSCDetId         (CSCDetId id)                 { csc_DetId         = id;        }
-    void SetRPCDetId         (RPCDetId id)                 { rpc_DetId         = id;        }
-    void SetCSCLCTDigi       (CSCCorrelatedLCTDigi digi)   { csc_LCTDigi       = digi;      }
-    void SetRPCDigi          (RPCDigi digi)                { rpc_Digi          = digi;      }
-    
-    CSCDetId CSC_DetId                          () const { return csc_DetId;    }
-    RPCDetId RPC_DetId                          () const { return rpc_DetId;    }
-    CSCCorrelatedLCTDigi CSC_LCTDigi            () const { return csc_LCTDigi;  }
-    RPCDigi RPC_Digi                            () const { return rpc_Digi;  }
-    const CSCDetId * PtrCSC_DetId               () const { return &csc_DetId;   }
-    const RPCDetId * PtrRPC_DetId               () const { return &rpc_DetId;   }
-    const CSCCorrelatedLCTDigi * PtrCSC_LCTDigi () const { return &csc_LCTDigi; }
-    const RPCDigi * PtrRPC_Digi                 () const { return &rpc_Digi; }
+    // DetId
+    uint32_t detId;
 
-    void set_endcap         (int  bits) { endcap        = bits; }
-    void set_station        (int  bits) { station       = bits; }
-    void set_ring           (int  bits) { ring          = bits; }
-    void set_sector         (int  bits) { sector        = bits; }
-    void set_sector_index   (int  bits) { sector_index  = bits; }
-    void set_subsector      (int  bits) { subsector     = bits; }
-    void set_chamber        (int  bits) { chamber       = bits; }
-    void set_csc_ID         (int  bits) { csc_ID        = bits; }
-    void set_roll           (int  bits) { roll          = bits; }
-    void set_rpc_layer      (int  bits) { rpc_layer     = bits; }
-    void set_neighbor       (int  bits) { neighbor      = bits; }
-    void set_mpc_link       (int  bits) { mpc_link      = bits; }
-    void set_wire           (int  bits) { wire          = bits; }
-    void set_strip          (int  bits) { strip         = bits; }
-    void set_strip_hi       (int  bits) { strip_hi      = bits; }
-    void set_strip_low      (int  bits) { strip_low     = bits; }
-    void set_track_num      (int  bits) { track_num     = bits; }
-    void set_quality        (int  bits) { quality       = bits; }
-    void set_pattern        (int  bits) { pattern       = bits; }
-    void set_bend           (int  bits) { bend          = bits; }
-    void set_valid          (int  bits) { valid         = bits; }
-    void set_sync_err       (int  bits) { sync_err      = bits; }
-    void set_bc0            (int  bits) { bc0           = bits; }
-    void set_bx             (int  bits) { bx            = bits; }
-    void set_stub_num       (int  bits) { stub_num      = bits; }
-    void set_is_CSC_hit     (int  bits) { is_CSC_hit    = bits; }
-    void set_is_RPC_hit     (int  bits) { is_RPC_hit    = bits; }
+    int16_t endcap;
+    int16_t station;
+    int16_t ring;
+    int16_t roll;
+    int16_t chamber;
+    int16_t sector;
+    int16_t subsector;
+    int16_t csc_ID;
+    int16_t cscn_ID;
 
-    int   Endcap         ()  const { return endcap   ;      }
-    int   Station        ()  const { return station  ;      }
-    int   Ring           ()  const { return ring     ;      }
-    int   Sector         ()  const { return sector   ;      }
-    int   Sector_index   ()  const { return sector_index;   }
-    int   Subsector      ()  const { return subsector;      }
-    int   Chamber        ()  const { return chamber  ;      }
-    int   CSC_ID         ()  const { return csc_ID   ;      }
-    int   Roll           ()  const { return roll     ;      }
-    int   RPC_layer      ()  const { return rpc_layer;      }
-    int   Neighbor       ()  const { return neighbor ;      }
-    int   MPC_link       ()  const { return mpc_link ;      }
-    int   Wire           ()  const { return wire     ;      }
-    int   Strip          ()  const { return strip    ;      }
-    int   Strip_hi       ()  const { return strip_hi ;      }
-    int   Strip_low      ()  const { return strip_low;      }
-    int   Track_num      ()  const { return track_num;      }
-    int   Quality        ()  const { return quality  ;      }
-    int   Pattern        ()  const { return pattern  ;      }
-    int   Bend           ()  const { return bend     ;      }
-    int   Valid          ()  const { return valid    ;      }
-    int   Sync_err       ()  const { return sync_err ;      }
-    int   BC0            ()  const { return bc0      ;      }
-    int   BX             ()  const { return bx       ;      }
-    int   Stub_num       ()  const { return stub_num ;      }
-    int   Is_CSC_hit     ()  const { return is_CSC_hit;     }
-    int   Is_RPC_hit     ()  const { return is_RPC_hit;     }
+    // BX
+    int16_t bx;
 
+    // Subsystem
+    int16_t subsystem;
 
-  private:
-    
-    CSCDetId csc_DetId;
-    RPCDetId rpc_DetId;
-    CSCCorrelatedLCTDigi csc_LCTDigi;
-    RPCDigi rpc_Digi;
-    
-    int   endcap;       // -1 or 1.  Filled in EMTFHit.cc from CSCDetId, modified
-    int   station;      //  1 -  4.  Filled in EMTFHit.cc from CSCDetId
-    int   ring;         //  1 -  3.  Filled in EMTFHit.cc from CSCDetId
-    int   sector;       //  1 -  6.  Filled in EMTFHit.cc from CSCDetId
-    int   sector_index; //  0 - 11.  0 - 5 for positive endcap, 6 - 11 for negative.  If a neighbor hit, set by the sector that received it, not the actual sector of the hit.
-    int   subsector;    //  1 -  2.  Filled in EMTFHit.cc or emulator using calc_subsector above
-    int   chamber;      //  1 - 36.  Filled in EMTFHit.cc from CSCDetId
-    int   csc_ID;       //  1 -  9.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi or emulator from CSCData
-    int   roll;         //  Sub-division of ring for RPC hits
-    int   rpc_layer;    //  Forward-backward bit for RPC hits
-    int   neighbor;     //  0 or 1.  Filled in EMTFBlockME.cc 
-    int   mpc_link;     //  1 -  3.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   wire;         //  1 -  ?.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   strip;        //  1 -  ?.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   strip_hi;     //  Highest strip number in an RPC cluster
-    int   strip_low;    //  Lowest strip number in an RPC cluster
-    int   track_num;    //  ? -  ?.  Filled in emulator from CSCData 
-    int   quality;      //  0 - 15.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   pattern;      //  0 - 10.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   bend;         //  0 or 1.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   valid;        //  0 or 1.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   sync_err;     //  0 or 1.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   bc0;          //  0 or 1.
-    int   bx;           //  -3 - 3.  Filled in EMTFHit.cc from CSCCorrelatedLCTDigi
-    int   stub_num;     //  0 or 1.
-    int   is_CSC_hit;   //  0 or 1.  Filled in EMTFHit.cc
-    int   is_RPC_hit;   //  0 or 1.  Filled in EMTFHit.cc
+    // Labels for the PrimitiveConversion processing unit
+    uint16_t pc_sector;
+    uint16_t pc_station;
+    uint16_t pc_chamber;
+    uint16_t pc_segment;
 
-  }; // End of class EMTFHit
-  
-  // Define a vector of EMTFHit
-  typedef std::vector<EMTFHit> EMTFHitCollection;
-  
-} // End of namespace l1t
+    // Input to PrimitiveConversion
+    uint16_t valid;
+    uint16_t strip;
+    uint16_t strip_low;
+    uint16_t strip_hi;
+    uint16_t wire;
+    uint16_t quality;
+    uint16_t pattern;
+    uint16_t bend;
+
+    // Output from PrimitiveConversion
+    uint16_t phi_fp;
+    uint16_t theta_fp;
+    uint16_t phzvl;
+    uint16_t ph_hit;
+    uint16_t zone_hit;
+    uint16_t zone_code;
+
+    // Variables used in PrimitiveMatching
+    uint16_t fs_segment;
+    uint16_t fs_zone_code;
+
+    // Variables used in BestTrackSelection
+    uint16_t bt_station;
+    uint16_t bt_segment;
+
+    // Coordinates
+    float phi_loc_deg;
+    float phi_glob_deg;
+    float theta_deg;
+    float eta;
+
+  };  // class EMTFHitExtra
+
+  // Define a vector of EMTFHitExtra
+  typedef std::vector<EMTFHitExtra> EMTFHitExtraCollection;
+
+}
 
 #endif
