--- L1TMuon/src/EMTFHitExtra.cc
@@ -1,6 +1,0 @@
-#include "DataFormatsSep2016/L1TMuon/interface/EMTFHitExtra.h"
-
-#include <cassert>
-#include <type_traits>
-static_assert(std::is_trivially_copyable<L1TMuonEndCap::EMTFHitExtra>::value, "EMTFHitExtra is not trivially copyable");
-
