--- L1TMuon/src/EMTFPtLUTData.cc
@@ -1,6 +1,0 @@
-#include "DataFormatsSep2016/L1TMuon/interface/EMTFPtLUTData.h"
-
-#include <cassert>
-#include <type_traits>
-static_assert(std::is_trivial<L1TMuonEndCap::EMTFPtLUTData>::value, "EMTFPtLUTData is not trivial");  // no default constructor
-
