--- L1TMuon/src/EMTFRoadExtra.cc
@@ -1,6 +1,0 @@
-#include "DataFormatsSep2016/L1TMuon/interface/EMTFRoadExtra.h"
-
-#include <cassert>
-#include <type_traits>
-static_assert(std::is_trivially_copyable<L1TMuonEndCap::EMTFRoadExtra>::value, "EMTFRoadExtra is not trivially copyable");
-
