--- L1TMuon/src/EMTFTrackExtra.cc
@@ -1,6 +1,0 @@
-#include "DataFormatsSep2016/L1TMuon/interface/EMTFTrackExtra.h"
-
-#include <cassert>
-#include <type_traits>
-//static_assert(std::is_trivially_copyable<L1TMuonEndCap::EMTFTrackExtra>::value, "EMTFTrackExtra is not trivially copyable");  // is false
-
